chore: add precommit hooks (#43)
* feat(lefthook) add precommit-hook to pythia to avoid format error in CI

* rebase(master) signing old commits

* fix(lefthook) biome format should run after sqlx prepare

- Change add-precommit-hook.sh to prepare.sh
- Join sqlx and biome commands into one command
sqlx-prepapre-and-biome-format

* adjust(lefthook): add prepapre command to pnpm to run prepare.sh script

* fix(lefthook) remove pnpm install command in prepare.sh

- Adjust README.md, pnpm install should be typed manually

* fix(README.md) remove lefthook doc since it automatically run after pnpm install

* fix(lefthook) add linting hook, refactor sqlx hook

* fix(lefthook) add new generated sqlx query in .sqlx folder into git after reformat

* fix(ci) make biome ignore .sqlx folder

+++ scripts/sqlx-prepare.sh
@@ -2,10 +2,25 @@
 
 DATABASE_URL="postgres://postgres:postgres@127.0.0.1:5432/pythia"
 
-docker compose down --volumes postgres
-docker compose up -d --wait postgres
+# Check if postgres container is already running
+if ! docker ps | grep -q postgres; then
+  echo "Starting postgres container..."
+  docker compose down --volumes postgres
+  docker compose up -d --wait postgres
+else
+  echo "Postgres container is already running."
+fi
 
-docker compose exec postgres psql -U postgres -c "CREATE DATABASE pythia;"
+# Check if the database exists, create it if it doesn't
+if ! docker compose exec postgres psql -U postgres -lqt | cut -d \| -f 1 | grep -qw pythia; then
+  echo "Creating pythia database..."
+  docker compose exec postgres psql -U postgres -c "CREATE DATABASE pythia;"
+else
+  echo "Database 'pythia' already exists."
+fi
 
+echo "Running migrations..."
 cargo sqlx migrate run --source packages/pythia/migrations --database-url $DATABASE_URL
+
+echo "Preparing SQLx..."
 cargo sqlx prepare --workspace --database-url $DATABASE_URL
